feature-languages: Localization helper scripts
Do the localization with the release build of the app.

init-localization: Copies LocBaml.exe to bin/Release. Run it if it's not there.

create-csv: Creates a csv file with all the texts to translate in it. You just have to replace the texts inside the csv file to the translated version.  First parameter: the name of the output csv file. Be careful! It overrides the file if it exists!

generate-localized-resource: Generates a translated resource out of the csv file you provide. First parameter: the name of the csv file. Second parameter: the language code (e.g hu-HU).